Grief Project
I added the objects in each stage and it adds the comedy I was missing. It also makes it more gamified and absurd, which I like. Everything is ''done''. I just have to fix issues, make everything transition smoothly and add sounds. Will finish everything before the meeting on Friday.